add static open_id provider
